requirements.txt
aiohttp==3.10.10
tiktoken==0.8.0
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.51.2
requests==2.32.3
tqdm==4.66.5
rank-bm25==0.2.2
tabulate==0.9.0
docling==2.14.0
pyprojroot==0.3.0
PyPDF2==3.0.1
pandas==2.2.3
faiss-cpu==1.9.0.post1
langchain==0.3.3
json_repair==0.35.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
click==8.1.7
httpx==0.27.2
sentence-transformers==3.4.1
einops==0.8.1
